+++ html-react-renderer.js
@@ -1,4 +1,4 @@
-const { renderToString } = require('react-dom/server')
+const { renderToString } = require(`${process.cwd()}/node_modules/react-dom/server`)
 
 module.exports = function htmlReactRenderer(template) {
   return '<!DOCTYPE html>\n'+renderToString(template)
