+++ injects-server.js
@@ -1,9 +1,7 @@
 import * as React from 'react'
 import * as fs from 'fs'
-import * as path from 'path'
-import * as stream from 'stream'
 
-export { fs, path, cx, stream, React }
+export { fs, cx, React }
 
 function cx(...args) {
   return args.filter(Boolean).join(' ')
