+++ plugins/universalComponent.js
