+++ renderers.js
@@ -1,0 +1,5 @@
+
+module.exports = {
+  json: __dirname+'/json-renderer.js',
+  html: __dirname+'/html-react-renderer.js',
+}
